+++ htk/accessors.go
@@ -1,4 +1,4 @@
-package golab
+package htk
 
 ///
 /// LAB ACCESSORS AND MUTATORS
